+++ src/main.ts
@@ -99,13 +99,12 @@ import {
   impFile,
   impProto,
   maybeAddComment,
-  nullOrUndefined,
   maybePrefixPackage,
+  nullOrUndefined,
   safeAccessor,
-  withOrMaybeCheckIsNull,
   withAndMaybeCheckIsNotNull,
   withOrMaybeCheckIsNotNull,
-  withAndMaybeCheckIsNull,
+  withOrMaybeCheckIsNull,
 } from "./utils";
 import { visit, visitServices } from "./visit";
 
@@ -497,7 +496,6 @@ function makeLongUtils(options: Options, bytes: ReturnType<typeof makeByteUtils>
     `,
   );
 
-  // TODO This is unused?
   const numberToLong = conditionalOutput(
     "numberToLong",
     code`
@@ -531,6 +529,9 @@ function makeLongUtils(options: Options, bytes: ReturnType<typeof makeByteUtils>
       function longToNumber(long: ${Long}): number {
         if (long.gt(${bytes.globalThis}.Number.MAX_SAFE_INTEGER)) {
           throw new ${bytes.globalThis}.Error("Value is larger than Number.MAX_SAFE_INTEGER")
+        }
+        if (long.lt(${bytes.globalThis}.Number.MIN_SAFE_INTEGER)) {
+          throw new ${bytes.globalThis}.Error("Value is smaller than Number.MIN_SAFE_INTEGER")
         }
         return long.toNumber();
       }
@@ -743,7 +744,7 @@ function makeTimestampMethods(
   let seconds: string | Code = "Math.trunc(date.getTime() / 1_000)";
   let toNumberCode: string | Code = "t.seconds";
   const makeToNumberCode = (methodCall: string) =>
-    `t.seconds${options.useOptionals === "all" ? "?" : ""}.${methodCall}`;
+    `t.seconds${options.useOptionals === "all" || options.noDefaultsForOptionals ? "?" : ""}.${methodCall}`;
 
   if (options.forceLong === LongOption.LONG) {
     toNumberCode = makeToNumberCode("toNumber()");
@@ -1249,10 +1250,21 @@ function generateDecode(ctx: Context, fullName: string, messageDesc: DescriptorP
               ${messageProperty} = ${generateMapType ? "new Map()" : "{}"};
             }`
           : "";
+
+        const ifValueCheck = `${varName}.value !== undefined ${withAndMaybeCheckIsNotNull(
+          options,
+          `${varName}.value`,
+        )}`;
+        const maybeIfKeyCheck = `${
+          options.noDefaultsForOptionals
+            ? ` && ${varName}.key !== undefined ${withAndMaybeCheckIsNotNull(options, `${varName}.key`)}`
+            : ""
+        }`;
+
         chunks.push(code`
           ${tagCheck}
           const ${varName} = ${readSnippet};
-          if (${varName}.value !== undefined ${withAndMaybeCheckIsNotNull(options, `${varName}.value`)}) {
+          if (${ifValueCheck}${maybeIfKeyCheck}) {
             ${initializerSnippet}
             ${valueSetterSnippet};
           }